Da ich doof bin, habe ich natürlich Gebinde vergessen. Gebinde kann nun auch safe gelöscht werden

+++ src/app/Einkauf/Bundle/Repository/DeleteSettingsRepository.php
@@ -8,7 +8,12 @@ class DeleteSettingsRepository
 
     public function deleteById($id): bool
     {
+        $db = new Database('EinkaufArticle');
+        if ($db->update(['bundle_id'])->where("bundle_id", "=", ":id")->args([':bundle_id' => 1, ':id' => $id])->run()) {
         $db = new Database('EinkaufBundle');
         return $db->delete()->where("id", "=",":id")->args([':id' => $id])->run();
+        } else {
+            return false;
+        }
     }
 }
